\name{AnalysisOfClassifier}
\alias{AnalysisOfClassifier}
\title{ measure a classifier's performance. }
\usage{
AnalysisOfClassifier(TrueClassification, CLS)
}
\arguments{
\item{TrueClassification}{vector of integer numbers representing the TRUE classification}
\item{CLS}{vector of integer numbers representing the CLASSIFIER'S classification}
}
\value{UniqueClasses           the different classes in TrueClassification}
\description{
 measure a classifier's performance. }

