\name{dbt.classifiers-package}
\alias{dbt.classifiers-package}
\alias{dbt.classifiers}
\docType{package}
\title{
classifiers 
}
\description{
A collecion of classifiers
}
\details{
\tabular{ll}{
Package: \tab dbt.classifiers\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-08-33\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\keyword{ classifiers }

